requirements.txt
flask
sentence-transformers
pinecone
google-generativeai
python-dotenv